feat: 나이에 따른 요금 할인을 적용한다.

+++ src/main/java/wooteco/subway/domain/Path.java
@@ -6,9 +6,9 @@ import java.util.List;
 public class Path {
     private final List<Station> stations;
     private final int distance;
-    private final int fare;
+    private final double fare;
 
-    public Path(final List<Station> stations, final int distance, final int fare) {
+    public Path(final List<Station> stations, final int distance, final double fare) {
         this.stations = new LinkedList<>(stations);
         this.distance = distance;
         this.fare = fare;
@@ -22,7 +22,7 @@ public class Path {
         return distance;
     }
 
-    public int getFare() {
+    public double getFare() {
         return fare;
     }
 }
